# perf: more elegant solution to mem dump using unsafe and generics