--- a/ecc/bls12-378/fr/pedersen/pedersen.go
+++ b/ecc/bls12-378/fr/pedersen/pedersen.go
@@ -30,8 +30,8 @@ import (
 
 // ProvingKey for committing and proofs of knowledge
 type ProvingKey struct {
-	basis         []curve.G1Affine
-	basisExpSigma []curve.G1Affine
+	Basis         []curve.G1Affine
+	BasisExpSigma []curve.G1Affine
 }
 
 type VerifyingKey struct {
@@ -74,17 +74,17 @@ func Setup(bases ...[]curve.G1Affine) (pk []ProvingKey, vk VerifyingKey, err err
 
 	pk = make([]ProvingKey, len(bases))
 	for i := range bases {
-		pk[i].basisExpSigma = make([]curve.G1Affine, len(bases[i]))
+		pk[i].BasisExpSigma = make([]curve.G1Affine, len(bases[i]))
 		for j := range bases[i] {
-			pk[i].basisExpSigma[j].ScalarMultiplication(&bases[i][j], sigma)
+			pk[i].BasisExpSigma[j].ScalarMultiplication(&bases[i][j], sigma)
 		}
-		pk[i].basis = bases[i]
+		pk[i].Basis = bases[i]
 	}
 	return
 }
 
 func (pk *ProvingKey) ProveKnowledge(values []fr.Element) (pok curve.G1Affine, err error) {
-	if len(values) != len(pk.basis) {
+	if len(values) != len(pk.Basis) {
 		err = fmt.Errorf("must have as many values as basis elements")
 		return
 	}
@@ -95,13 +95,13 @@ func (pk *ProvingKey) ProveKnowledge(values []fr.Element) (pok curve.G1Affine, e
 		NbTasks: 1, // TODO Experiment
 	}
 
-	_, err = pok.MultiExp(pk.basisExpSigma, values, config)
+	_, err = pok.MultiExp(pk.BasisExpSigma, values, config)
 	return
 }
 
 func (pk *ProvingKey) Commit(values []fr.Element) (commitment curve.G1Affine, err error) {
 
-	if len(values) != len(pk.basis) {
+	if len(values) != len(pk.Basis) {
 		err = fmt.Errorf("must have as many values as basis elements")
 		return
 	}
@@ -111,7 +111,7 @@ func (pk *ProvingKey) Commit(values []fr.Element) (commitment curve.G1Affine, er
 	config := ecc.MultiExpConfig{
 		NbTasks: 1,
 	}
-	_, err = commitment.MultiExp(pk.basis, values, config)
+	_, err = commitment.MultiExp(pk.Basis, values, config)
 
 	return
 }
@@ -131,7 +131,7 @@ func BatchProve(pk []ProvingKey, values [][]fr.Element, fiatshamirSeeds ...[]byt
 
 	offset := 0
 	for i := range pk {
-		if len(values[i]) != len(pk[i].basis) {
+		if len(values[i]) != len(pk[i].Basis) {
 			err = fmt.Errorf("must have as many values as basis elements")
 			return
 		}
@@ -147,14 +147,14 @@ func BatchProve(pk []ProvingKey, values [][]fr.Element, fiatshamirSeeds ...[]byt
 	scaledValues := make([]fr.Element, offset)
 	basis := make([]curve.G1Affine, offset)
 
-	copy(basis, pk[0].basisExpSigma)
+	copy(basis, pk[0].BasisExpSigma)
 	copy(scaledValues, values[0])
 
 	offset = len(values[0])
 	rI := r
 	for i := 1; i < len(pk); i++ {
-		copy(basis[offset:], pk[i].basisExpSigma)
-		for j := range pk[i].basis {
+		copy(basis[offset:], pk[i].BasisExpSigma)
+		for j := range pk[i].Basis {
 			scaledValues[offset].Mul(&values[i][j], &rI)
 			offset++
 		}
@@ -245,11 +245,11 @@ func getChallenge(fiatshamirSeeds [][]byte) (r fr.Element, err error) {
 // Marshal
 
 func (pk *ProvingKey) writeTo(enc *curve.Encoder) (int64, error) {
-	if err := enc.Encode(pk.basis); err != nil {
+	if err := enc.Encode(pk.Basis); err != nil {
 		return enc.BytesWritten(), err
 	}
 
-	err := enc.Encode(pk.basisExpSigma)
+	err := enc.Encode(pk.BasisExpSigma)
 
 	return enc.BytesWritten(), err
 }
@@ -265,14 +265,14 @@ func (pk *ProvingKey) WriteRawTo(w io.Writer) (int64, error) {
 func (pk *ProvingKey) ReadFrom(r io.Reader) (int64, error) {
 	dec := curve.NewDecoder(r)
 
-	if err := dec.Decode(&pk.basis); err != nil {
+	if err := dec.Decode(&pk.Basis); err != nil {
 		return dec.BytesRead(), err
 	}
-	if err := dec.Decode(&pk.basisExpSigma); err != nil {
+	if err := dec.Decode(&pk.BasisExpSigma); err != nil {
 		return dec.BytesRead(), err
 	}
 
-	if cL, pL := len(pk.basis), len(pk.basisExpSigma); cL != pL {
+	if cL, pL := len(pk.Basis), len(pk.BasisExpSigma); cL != pL {
 		return dec.BytesRead(), fmt.Errorf("commitment basis size (%d) doesn't match proof basis size (%d)", cL, pL)
 	}