--- a/ecc/bls12-378/fr/pedersen/pedersen_test.go
+++ b/ecc/bls12-378/fr/pedersen/pedersen_test.go
@@ -166,8 +166,8 @@ func TestCommitFiveElements(t *testing.T) {
 
 func TestMarshal(t *testing.T) {
 	var pk ProvingKey
-	pk.basisExpSigma = randomG1Slice(t, 5)
-	pk.basis = randomG1Slice(t, 5)
+	pk.BasisExpSigma = randomG1Slice(t, 5)
+	pk.Basis = randomG1Slice(t, 5)
 
 	var (
 		vk  VerifyingKey